requirements.txt
paho-mqtt
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib 
feedparser